+++ tests/functional/runtime_api/api_runtime_config_alignment_smoke_test.go
@@ -256,7 +256,6 @@ func assertRuntimeConfigAlignmentFinalState(
 	assertRuntimeConfigAlignmentDispatchHistory(t, engineState.DispatchHistory)
 	assertRuntimeConfigAlignmentCompleteTokenPayload(t, engineState.Marking.Tokens)
 	assertRuntimeConfigAlignmentEventHistory(t, server)
-	assertRuntimeConfigAlignmentTopologyProjection(t, dir)
 }
 
 type runtimeConfigAlignmentSummary struct {
